pyproject.toml
[project]
name = "fluentian-interns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.4",
    "pymongo>=4.16.0",
    "pytelegrambotapi>=4.32.0",
    "python-dotenv>=1.2.2",
]
